Migrate to submodules: Checkpointer, ExecBlock
